+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
++ core: Allow changing the priority of event loops of components
 + core: `add_action` of `RelmActionGroup` now takes a reference to a `RelmAction` as a parameter
 + macros: `parse_with_path`, `update_stream`, `inject_view_code` and `generate_tokens` take references for some of their parameters